refactor(server): move default-branch validation query to Tuist.Tests
test_case_ids_with_successful_default_branch_run is a plain existence
lookup over test_case_runs ("which of these test cases have a passing,
non-flaky run on the default branch"), consumed by automation gating. It
is not an aggregation, metric, or chart, so it belongs in Tuist.Tests
("interacting with test runs") rather than Tuist.Tests.Analytics. It now
sits beside the existing branch-presence helpers that determine whether a
test case is new on the default branch. The unit test moves with it.
