+++ server/lib/tuist/tests/analytics.ex
@@ -1263,29 +1263,6 @@ defmodule Tuist.Tests.Analytics do
   end
 
   @doc """
-  Given a list of test case ids, returns the subset that has at least one
-  successful, non-flaky run on the project's default branch. A test case with
-  no such run has never been validated on the trusted branch (for example, a
-  brand-new test that has only ever run on a pull-request branch) and should
-  not be eligible for automated quarantine.
-  """
-  def test_case_ids_with_successful_default_branch_run(_project_id, [], _default_branch), do: []
-
-  def test_case_ids_with_successful_default_branch_run(project_id, test_case_ids, default_branch) do
-    ClickHouseRepo.all(
-      from(tcr in TestCaseRun,
-        where: tcr.project_id == ^project_id,
-        where: tcr.test_case_id in ^test_case_ids,
-        where: tcr.git_branch == ^default_branch,
-        where: fragment("? = 'success'", tcr.status),
-        where: tcr.is_flaky == false,
-        distinct: true,
-        select: tcr.test_case_id
-      )
-    )
-  end
-
-  @doc """
   Gets analytics for a specific test case by its UUID including total runs, failed runs, and average duration.
   """
   def test_case_analytics_by_id(test_case_id, _opts \\ []) do
